+++ Game/Modules/Actions/Effects/HotkeyExpansion.cs
@@ -6,18 +6,18 @@ namespace Game.Modules.Actions.Effects;
 /// <summary>Write surface for permanently unlocking more Expansion hotkey slots. Method named Apply, not Grant, so it doesn't collide with HotkeyExpansionGrant's own name -- the entry is a noun (a grant), this is the verb performed on it, the same split every other *Effects write-surface uses (see StatModifierEffects.Apply).</summary>
 public static class HotkeyExpansion
 {
-    public const short MaxUnlockedSlots = 20;
+    public const byte MaxUnlockedSlots = 20;
 
-    public static void Apply(PackedComponentPool<HotkeyExpansionUnlockComponent> hotkeyExpansionUnlocks, int entityId, short amount)
+    public static void Apply(PackedComponentPool<HotkeyExpansionUnlockComponent> hotkeyExpansionUnlocks, int entityId, byte amount)
     {
         if (!hotkeyExpansionUnlocks.Has(entityId))
         {
             return;
         }
 
-        hotkeyExpansionUnlocks.TryUpdate(entityId, amount, static (ref HotkeyExpansionUnlockComponent unlock, short grantAmount) =>
+        hotkeyExpansionUnlocks.TryUpdate(entityId, amount, static (ref HotkeyExpansionUnlockComponent unlock, byte grantAmount) =>
         {
-            unlock.UnlockedSlotCount = (short)Math.Min(MaxUnlockedSlots, unlock.UnlockedSlotCount + grantAmount);
+            unlock.UnlockedSlotCount = (byte)Math.Min(MaxUnlockedSlots, unlock.UnlockedSlotCount + grantAmount);
         });
     }
 }
