+++ Game/Modules/Poison/Components/PoisonTimerComponent.cs
@@ -8,13 +8,13 @@ namespace Game.Modules.Poison.Components;
 /// Present on an entity only while it currently has at least one Poison stack -- added on the
 /// 0-to-1 stack transition, removed only when RemainingDurationTicks reaches 0.
 /// </summary>
-public struct PoisonTimerComponent(int framesUntilNextTick, int stackCount, int remainingDurationTicks, StatusEffectSource source) : ITickCountdown, IStatusEffectStackCount
+public struct PoisonTimerComponent(ushort framesUntilNextTick, byte stackCount, ushort remainingDurationTicks, StatusEffectSource source) : ITickCountdown, IStatusEffectStackCount
 {
-    public int FramesUntilNextTick { get; set; } = framesUntilNextTick;
+    public ushort FramesUntilNextTick { get; set; } = framesUntilNextTick;
 
-    public int StackCount { get; set; } = stackCount;
+    public byte StackCount { get; set; } = stackCount;
 
-    public int RemainingDurationTicks { get; set; } = remainingDurationTicks;
+    public ushort RemainingDurationTicks { get; set; } = remainingDurationTicks;
 
     public StatusEffectSource Source { get; set; } = source;
