+++ Game/Modules/StatusEffects/IStatusEffectStackCount.cs
@@ -1,12 +1,9 @@
 namespace Game.Modules.StatusEffects;
 
-/// <summary>
-/// Implemented by a status effect's own timer/tracking component (BurningTimerComponent,
-/// PoisonTimerComponent, ...) that caches how many stacks an entity currently holds. Lets
-/// TimerBasedAuraApplier&lt;T&gt; read that count generically, without hardcoding any one
-/// effect's own component type.
-/// </summary>
+/// <summary>Represents a component that tracks the stack count of a status effect.</summary>
+/// <remarks>Maximum of 255 stacks per status effect, though each effect can define a lower limit.</remarks>
+/// <cleanupVersion>1</cleanupVersion>
 public interface IStatusEffectStackCount
 {
-    int StackCount { get; }
+    byte StackCount { get; }
 }
